fix: use canonical user profile route
