+++ src/components/UserBadge.tsx
@@ -33,7 +33,7 @@ export function UserBadge({
     hasOwnProperty(user, "name") && typeof user.name === "string" ? user.name.trim() : undefined;
   const displayName = user?.displayName?.trim() || userName || null;
   const handle = user?.handle ?? fallbackHandle ?? null;
-  const href = user?.handle ? `/p/${encodeURIComponent(user.handle)}` : null;
+  const href = user?.handle ? `/user/${encodeURIComponent(user.handle)}` : null;
   const label = handle ? `@${handle}` : "user";
   const image = user?.image ?? null;
   const hasUsefulName =
