+++ src/routes/u/$handle.tsx
@@ -3,7 +3,7 @@ import { createFileRoute, redirect } from "@tanstack/react-router";
 export const Route = createFileRoute("/u/$handle")({
   beforeLoad: ({ params }) => {
     throw redirect({
-      to: "/p/$handle",
+      to: "/user/$handle",
       params: { handle: params.handle },
       replace: true,
     });
